+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    - Moved warning for no cleaning of JRO ISR data to preprocess
    - Standardized the Instrument method kwarg defaults
    - Added 'site' tag to the GNSS TEC Instrument
+   - Added support for varied use of `two_digit_year_break` to 
+     `methods.general.list_remote_files`
+   - Implemented `two_digit_year_break` support for `vtec` GNSS TEC Instrument
 - Documentation
    - Added examples for JRO and GNSS data
    - Improved the docstring style